% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggsnippets.R
\docType{data}
\name{braemardata}
\alias{braemardata}
\title{Long term historical climate data for Breamar, UK}
\format{A data frame with 669 rows and 6 variables.}
\source{
\url{http://www.metoffice.gov.uk/pub/data/weather/uk/climate/stationdata/braemardata.txt}
}
\usage{
data(braemardata)
}
\description{
A dataset containing monthly maximum, minimum temperature, rainfall and sunlight hours
}
\details{
\itemize{
  \item year. year of record
  \item month. month of record
  \item max_temp. maximum temperature in this month
  \item min_temp. minimum temperature in this month
  \item rain_mm. total rainfall of the month
  \item sun_hours. total sunlight hours in the month
}
}
\keyword{datasets}

