% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggsnippets.R
\name{breamardata}
\alias{breamardata}
\title{Climatic data used in the package's examples.}
\format{A data frame with 669 rows and 6 variables}
\description{
This dataset contans min and maximum temperature, rainfall and
the amount of sunlight hours for each day.
}
\details{
\itemize{
 \item year. The year of the sample
 \item month. The month the sample was taken
 \item max_temp. The maximum temperature
 \item min_temp. The minimum temperature
 \item rain_mm. The rainfall for the month.
 \item sun_hours. The amount of sunlight hours for the month.
 }
}

