% Generated by roxygen2 (4.0.2): do not edit by hand
\name{facet_wrap_labeller}
\alias{facet_wrap_labeller}
\title{A function to allow for expressions in ggplot facet titles.}
\usage{
facet_wrap_labeller(gg.plot, labels = NULL)
}
\arguments{
\item{gg.plot}{a ggplot object}
}
\value{
a ggplot object with new facet titles.
}
\description{

}
\examples{
x_cars <- data.frame(cars, variable = sample(letters[1:5], 50, replace = TRUE))
plt1 <- ggplot() + geom_point(aes(speed, dist), x_cars) + facet_wrap(~variable, ncol=1)
plt
facet_wrap_labeller(plt1)
facet_wrap_labeller(plt1, labels = c(expression(a^2), expression(b[1]), expression(frac(c,2)), "d", expression(infinity)))
}
\references{
\url{http://stackoverflow.com/questions/11979017}
}
\seealso{
plotmath
}

