% Generated by roxygen2 (4.0.2): do not edit by hand
\name{g_legend}
\alias{g_legend}
\title{Extract legend from ggplot geom object}
\usage{
g_legend(gg.plot)
}
\arguments{
\item{gg.plot}{A ggplot object}
}
\value{
The legend as a grob object from the plot
}
\description{
Lets you extract a ggplot legend to be used elsewhere.
}
\details{

}
\examples{
## Create a plot with a legend
plt1 <- ggplot() + geom_boxplot(aes(factor(gear), mpg, fill=factor(gear)), mtcars)
## Extract the legend from the plot.
plt_legend <- g_legend(plt1)
## Create a second boxplot with the same fill
plt2 <- ggplot() + geom_boxplot(aes(factor(gear), disp, fill=factor(gear)), mtcars)
## Supress the legends from both plots
plt1 <- plt1 +  theme(legend.position="none")
plt2 <- plt2 +  theme(legend.position="none")
arrangeGrob(plt1, plt2, plt_legend, ncol=3)
## Or a slightly nicer example
arrangeGrob(arrangeGrob(plt1, plt2, ncol=1), plt_legend, ncol=2, widths=c(0.8,0.2))
}
\references{
Obtained from \url{http://stackoverflow.com/questions/11883844}
}

