% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facet_wrap_labeller.R
\name{facet_wrap_labeller}
\alias{facet_wrap_labeller}
\title{A function to allow for expressions in ggplot facet titles.}
\usage{
facet_wrap_labeller(gg.plot, labels = NULL)
}
\arguments{
\item{gg.plot}{a ggplot object}

\item{labels}{labels for each facet.}
}
\value{
a ggplot object with new facet titles.
}
\description{
Allows for the use of expressions with facet titles.
}
\details{
A function to label facetted ggplots

This function alters the lables of a facetted ggplot object. This is useful if you want
use expressions or have a simple way to mdify the labels of a facetted ggplot object.
}
\examples{
## Add a factor to the cars example data to facet the plots
library(ggplot2)
x_cars <- data.frame(cars, variable = sample(letters[1:5],
                                             50,
                                             replace = TRUE))
## Create the plot with facets
plt <- ggplot() + geom_point(aes(speed, dist), x_cars) +
       facet_wrap(~variable, ncol=1)
plt
## An example without labels - no real point in this.
library(gridExtra)
facet_wrap_labeller(plt)
## Adding expressions within the facet labels.
facet_wrap_labeller(plt, labels = c(expression(a^2),
                                   expression(b[1]),
                                   expression(frac(c,2)),
                                    "d",
                                     expression(infinity)))
}
\references{
\url{http://stackoverflow.com/questions/11979017}
}
\seealso{
plotmath
}

